chore(review): 自製部分全面審查治本修正 + repo 止血 (0.7.0.21)
依「完整審查自製 part」+「aall」對自製程式碼全面治本，經 test-phase0 + preflight 0 FAIL。

Rust:
- launcher 缺檔/非法二進位不再偽造成功，PE stub 改 STRAWWU_SMOKE 顯式開啟，訊息標 mode=simulated
- strawwu-nt loader/pe RVA/size 改 u64 防 u32 溢位
- app-registry 原子寫入 + flock 跨程序鎖，log 改 serde_json 防注入
- desktop entry app-id 驗證 + 值跳脫；flatpak/apt 移除加 app-id/套件名驗證 + --

os-image:
- clone-ubuntu-base 修 dpkg-query 引號 + base ISO SHA256 校驗
- Secure Boot 統一 MOK 單軌；修 patch_boot_serial_console 誤改 vmlinuz-generic；fallback grub 條件化
- branding sed 收斂為特定標題不全域；laptop/device-proxy/icon-theme 補 prerm/postrm

kernel:
- ABI 單一來源＝ubuntu-base-target.json active.kernel_abi，移除硬編 fallback
- apt-get source 移除未 pin fallback → 版本不符硬失敗 + 抓取後 ABI 驗證
- .kernel-signing 誠實記錄簽核姿態

hub: sandbox:true + 導航守衛 + IPC 驗證 + escapeHtml 補跳脫引號

tests/git:
- 無 rootfs preflight 改回報 SKIPPED（非假 PASS）；CI 補 lrelease；Rust CLI 修 SIGPIPE panic
- 止血：untrack os-image/debs/*/output/*.deb 與 tests/apt-repo|ci/output（含測試 GPG 私鑰）+ gitignore

docs: README 版本/kernel/codename/crate 數同步 (0.7.0.21, 7.0.0-strawwu, resolute, 11 crates)

+++ components/strawwu-nt/src/loader.rs
@@ -58,8 +58,10 @@ impl PeLoader {
 
         self.state = LoaderState::Mapping;
 
+        // Widen to u64 before adding: virtual_address + size can exceed u32::MAX
+        // for a crafted/large PE and must not wrap around.
         let total_size = pe.sections.iter()
-            .map(|s| (s.virtual_address + s.virtual_size.max(s.raw_size)) as u64)
+            .map(|s| s.virtual_address as u64 + s.virtual_size.max(s.raw_size) as u64)
             .max()
             .unwrap_or(0x1000);
